+++ packages/storybook/src/nimble/mapping/icon/mapping-icon.stories.ts
@@ -27,9 +27,7 @@ export const iconMapping: StoryObj = {
         },
         icon: {
             control: false,
-            description: `The tag name of the Nimble icon to render, e.g. \`nimble-icon-check\`. Alternatively, set \`icon\` to \`undefined\` to render
-                no icon for the mapping while still providing a label to be used when grouping. Space will always be reserved for the icon so
-                that the text in cells and group rows associated with icon mappings will always be aligned.`
+            description: 'The tag name of the Nimble icon to render, e.g. `nimble-icon-check`.'
         },
         severity: {
             control: false,