+++ packages/storybook/src/nimble/select/select-matrix.stories.ts
@@ -39,6 +39,12 @@ const valueStates = [
 ] as const;
 type ValueState = (typeof valueStates)[number];
 
+const clearableStates = [
+    ['', false],
+    ['Clearable', true]
+] as const;
+type ClearableState = (typeof clearableStates)[number];
+
 const metadata: Meta = {
     title: 'Tests/Select',
     parameters: {
@@ -54,6 +60,7 @@ const component = (
     [appearanceName, appearance]: AppearanceState,
     [errorName, errorVisible, errorText]: ErrorState,
     [valueName, valueValue]: ValueState,
+    [clearableName, clearable]: ClearableState
 ): ViewTemplate => html`
     <div style="
         display: inline-flex;
@@ -62,11 +69,12 @@ const component = (
         font: var(${controlLabelFont.cssCustomProperty});
         color: var(${controlLabelFontColor.cssCustomProperty});"
     >
-        <label>${() => errorName} ${() => disabledName} ${() => appearanceName} ${() => valueName}</label>
+        <label>${() => errorName} ${() => disabledName} ${() => appearanceName} ${() => valueName} ${() => clearableName}</label>
         <${selectTag}
             ?error-visible="${() => errorVisible}"
             error-text="${() => errorText}"
             ?disabled="${() => disabled}"
+            ?clearable="${() => clearable}"
             appearance="${() => appearance}"
             style="width: 250px;"
         >
@@ -83,7 +91,8 @@ export const selectThemeMatrix: StoryFn = createMatrixThemeStory(
         disabledStates,
         appearanceStates,
         errorStates,
-        valueStates
+        valueStates,
+        clearableStates
     ])
 );